Rama lucas - Validacion para alta de planilla

+++ src/main/java/com/inmobiliaria/InmoGestion/servicio/InmuebleServicio.java
@@ -4,7 +4,6 @@ package com.inmobiliaria.InmoGestion.servicio;
 import com.inmobiliaria.InmoGestion.DTO.InmuebleDTO;
 import com.inmobiliaria.InmoGestion.modelo.*;
 import com.inmobiliaria.InmoGestion.repositorio.InmuebleRepositorio;
-import com.inmobiliaria.InmoGestion.repositorio.TipoOperacionRepositorio;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +20,8 @@ public class InmuebleServicio {
     private final ServiciosServicio serviciosServicio;
     private final AmbientesServicio ambientesServicio;
     private final ImagenServicio imagenServicio;
-
+    private final EstadoInmuebleServicio estadoInmuebleServicio;
+    private final PropietarioServicio propietarioServicio;
 
     @Transactional
     public Inmueble crear (InmuebleDTO dto){
@@ -56,6 +56,10 @@ public class InmuebleServicio {
             inmueble.setListaImagenes(imagenServicio.almacenarImagenes(dto.getImagenes()));
             inmueble.setEsAlquiler(dto.getEsAlquiler());
             inmueble.setEsVenta(dto.getEsVenta());
+            inmueble.setPrecioAlquiler(dto.getPrecioAlquiler());
+            inmueble.setPrecioVenta(dto.getPrecioVenta());
+            inmueble.setPropietario(propietarioServicio.obtenerPorId(dto.getPropietario()));
+            inmueble.setEstadoInmueble( (estadoInmuebleServicio.obtenerPorNombre("activo") != null ) ? estadoInmuebleServicio.obtenerPorNombre("activo") : null ); // Busca estado inmueble por id = 1 *- Para ello siempre el primer estadoInmueble a crear debe ser 'activo'-*
             return inmuebleRepositorio.save(inmueble);
         }catch (Exception e){
             e.printStackTrace();
@@ -99,6 +103,10 @@ public class InmuebleServicio {
                 if(dto.getImagenes() != null) inmueble.setListaImagenes(imagenServicio.almacenarImagenes(dto.getImagenes(), inmueble.getListaImagenes()));
                 inmueble.setEsAlquiler(dto.getEsAlquiler());
                 inmueble.setEsVenta(dto.getEsVenta());
+                inmueble.setPrecioAlquiler(dto.getPrecioAlquiler());
+                inmueble.setPrecioVenta(dto.getPrecioVenta());
+                inmueble.setPropietario(propietarioServicio.obtenerPorId(dto.getPropietario()));
+                inmueble.setEstadoInmueble(estadoInmuebleServicio.obtenerPorId(dto.getEstadoInmueble()));
                 return inmuebleRepositorio.save(inmueble);
             }else{
                 throw new RuntimeException("No se ha encontrado el registro");
@@ -132,7 +140,6 @@ public class InmuebleServicio {
     @Transactional
     public HashMap<String, String> eliminarPorId(Long id){
         try {
-            System.out.println("Ingreso al servicio eliminar");
             //CREO UN HASHMAP PARA DEVOLVER LA RESPUESTA
             HashMap<String, String> respuesta = new HashMap<>();
             Optional<Inmueble> inmuebleOptional = inmuebleRepositorio.findById(id); // BUSCO EL INMUEBLE EN LA BD
@@ -142,8 +149,12 @@ public class InmuebleServicio {
                 Inmueble inmueble = inmuebleOptional.get(); // OBTENGO EL OBJETO
                 List<String> listaImagenes = inmueble.getListaImagenes();   // PASO LOS NOMBRES DE LAS IMAGENES QUE TIENE VINCULADAS
                 inmuebleRepositorio.deleteById(id); //ELIMINO DE LA BD
-                if (!listaImagenes.isEmpty()) { // SI TIENE IMAGENES LAS ELIMINO
-                    imagenServicio.eliminarImagenes(listaImagenes);
+
+                boolean exists = inmuebleRepositorio.existsById(id);
+                if(!exists){
+                    if (!listaImagenes.isEmpty()) { // SI TIENE IMAGENES LAS ELIMINO
+                        imagenServicio.eliminarImagenes(listaImagenes);
+                    }
                 }
 
                 respuesta.put("mensaje", "Se ha eliminardo correctamente el registro");
@@ -158,5 +169,12 @@ public class InmuebleServicio {
         }
     }
 
+    @Transactional
+    public void cambiarEstado(Long inmuebleId, String estado) {
+        EstadoInmueble estadoInmueble = estadoInmuebleServicio.obtenerPorNombre(estado);
+        Inmueble inmueble = this.obtenerPorId(inmuebleId).orElse(null);
 
+        inmueble.setEstadoInmueble(estadoInmueble);
+        inmuebleRepositorio.save(inmueble);
+    }
 }
